+++ extensions/system-monitor/CHANGELOG.md
@@ -1,5 +1,9 @@
 # System Monitor Changelog
 
+## [Fix Temperature Polling] - {PR_MERGE_DATE}
+
+- Moved temperature sensor polling to a dedicated 3s interval to prevent stale readings
+
 ## [Added Menubar Pin-to-Display] - 2026-03-04
 
 - Click any stat in the menubar dropdown to pin it as persistent text next to the icon